fix: refine code
